pyproject.toml
[project]
name = "Reporte-del-tiempo"
version = "2.0.0"
description = "Una simple aplicación para generar reportes del tiempo."
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Rubén Rojas", email = "ruben.eduardo.uban@gmail.com" }
]
dependencies = [
  "flet==0.28.3",
  "pyjnius==1.6.1", # For native Android access
  "flet-audio==0.1.0"
]

[tool.flet]
org = "com.darkwolf261"
product = "Reporte del Tiempo"
company = "Rubén Rojas"
copyright = "Copyright (C) 2025 Rubén Rojas"

[tool.flet.app]
path = "src"

[tool.flet.android.permission]
"android.permission.POST_NOTIFICATIONS" = true
"android.permission.FOREGROUND_SERVICE" = true
"android.permission.SCHEDULE_EXACT_ALARM" = true
"android.permission.RECEIVE_BOOT_COMPLETED" = true

[tool.flet.android.services]
"AlarmService" = "src/service.py"

[tool.uv]
dev-dependencies = [
    "flet[all]==0.28.3",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
flet = {extras = ["all"], version = "0.28.3"}

[tool.flet.flutter.pubspec.dependency_overrides]
webview_flutter_android = "4.10.1"